+++ InformationProtection/functions/Get-MipFile.ps1
@@ -10,6 +10,11 @@
 	
 	.PARAMETER Path
 		Path to the file to scan.
+
+	.PARAMETER Session
+		MIP Session to use for the operation.
+		Overrides the use of the default session and would be used in situations when relabeling files from one tenant to another.
+		Use "New-MipSession" to create a standalone session object.
 	
 	.EXAMPLE
 		PS C:\> Get-MipFile -Path .\*
@@ -26,14 +31,19 @@
 	param (
 		[Parameter(Mandatory = $true, ValueFromPipeline = $true)]
 		[PSFFile]
-		$Path
+		$Path,
+
+		[InformationProtection.MipSession]
+		$Session
 	)
 	begin {
-		Assert-MIPConnection -Cmdlet $PSCmdlet
+		Assert-MIPConnection -Cmdlet $PSCmdlet -Session $Session
+		$sessionToUse = $script:_session
+		if ($Session.Context) { $sessionToUse = $Session}
 	}
 	process {
 		foreach ($filePath in $Path) {
-			[InformationProtection.File]::new($filePath)
+			[InformationProtection.File]::new($filePath, $sessionToUse)
 		}
 	}
 }