+++ InformationProtection/internal/scripts/variables.ps1
@@ -3,3 +3,6 @@
 	MIPSyncService = 'MIPSyncService'
 }
 $script:_serviceSelector = New-EntraServiceSelector -DefaultServices $script:_services
+
+# The current Runspace's default MIP Session
+$script:_session = $null