+++ library/InformationProtection/InformationProtection/File.cs
@@ -49,15 +49,19 @@ namespace InformationProtection
         /// </summary>
         public string LabelID { get; private set; }
 
+        private MipSession _Session;
+
         /// <summary>
         /// Create a new file object from a path
         /// </summary>
         /// <param name="Path">The path to the file. Can be PowerShell-relative, will be resolved.</param>
+        /// <param name="Session">The MIP Session used to perform labeling operations.</param>
         /// <exception cref="InvalidOperationException">Without connection, nothing can happen.</exception>
-        public File(PathFileSingleParameter Path)
+        public File(PathFileSingleParameter Path, MipSession Session)
         {
-            if (null == MipHost.Context)
-                throw new InvalidOperationException("Not yet connected! Call Authenticator.Authenticate first!");
+            if (null == Session.Context)
+                throw new InvalidOperationException("Not yet connected! Call Session.Authenticate first!");
+            _Session = Session;
             this.Path = Path;
             RefreshState();
         }
@@ -65,8 +69,9 @@ namespace InformationProtection
         /// Create a new file object from a path.
         /// </summary>
         /// <param name="Path">The path to the file. Can be PowerShell-relative, will be resolved.</param>
-        public File(object Path)
-            :this(new PathFileSingleParameter(Path))
+        /// <param name="Session">The MIP Session used to perform labeling operations.</param>
+        public File(object Path, MipSession Session)
+            :this(new PathFileSingleParameter(Path), Session)
         {
 
         }
@@ -77,7 +82,7 @@ namespace InformationProtection
         /// <exception cref="InvalidOperationException">Can only be called if connected, has a path and the file exists</exception>
         public void RefreshState()
         {
-            if (null == MipHost.Context)
+            if (null == _Session.Context)
                 throw new InvalidOperationException("Not yet connected! Call Authenticator.Authenticate first!");
 
             if (String.IsNullOrEmpty(Path))
@@ -86,7 +91,7 @@ namespace InformationProtection
             if (!System.IO.File.Exists(Path))
                 throw new InvalidOperationException($"Path does not exist: {Path}!");
 
-            Handler = Task.Run(async () => await MipHost.FileEngine.CreateFileHandlerAsync(Path, Path, true)).Result;
+            Handler = Task.Run(async () => await _Session.FileEngine.CreateFileHandlerAsync(Path, Path, true)).Result;
             Label = Handler.Label;
             if (Label == null)
                 return;
@@ -118,7 +123,7 @@ namespace InformationProtection
                 labelingOptions.JustificationMessage = Justification;
             }
 
-            Handler.SetLabel(MipHost.FileEngine.GetLabelById(LabelID), labelingOptions, new ProtectionSettings());
+            Handler.SetLabel(_Session.FileEngine.GetLabelById(LabelID), labelingOptions, new ProtectionSettings());
             var result = Task.Run(async () => await Handler.CommitAsync(Destination)).Result;
         }